aspire: Consider Aspire Capability when detecting AppHost
In addition to the `IsAspireHost` property we will now look at the
project capabilities to see if an `Aspire` capability is listed and if
treat the project as an AppHost project. This aligns `azd`'s behavior
with other tooling like Visual Studio which uses the project
capabilities to determine if the project is an App Host or not.

The .NET Team asked us to include this in our sniffing logic (but to
continue to check `IsAspireHost` as well).

Fixes #4364