tests: Use OIDC for Terraform in template-tests
The terraform legs in our template test pipelines have been broken
since we moved to ODIC and this change gets them working
again. Because we run the template tests inside a dev container, we
can't leverage the AzureCLI task to configure OIDC for `az`, so we
have to do things the hard way.

The AzureRM Terraform Provider does support OIDC auth an even has
knobs to configure the URL and token used when fetching the federated
token used to do the ODIC dance, but it expects a slightly different
protocol than what Azure Pipelines provides, so we now have a small
adapter program `tfoidc` which exposes the endpoint in the format that
the provider expects (it's the format spoken by GitHub Actions) which
proxies to the Azure Pipelines version.

In addition, to support local-exec scenarios where `az` is invoked,
`tfoidc` can manage calling `az login` and `az account set` as the
AzureCLI task would (refreshing as needed).

The template-test pipelines have been updated to build this binary and
place it in the root of the folder that is mounted into the dev
container when the template-test script is run and to launch `tfoidc`
and configure things as needed before running the `template-test.sh`
script.